Perform setup of LTIView once app registry is ready.
Rationale: Set up of LTIView involves importing User model, and as of
Django 1.9 it's no longer possible to import models inside the root package of an application.

Cf. https://docs.djangoproject.com/en/2.1/releases/1.9/#features-removed-in-1-9
