+++ lpd/__init__.py
@@ -1,0 +1,5 @@
+"""
+LPD app.
+"""
+
+default_app_config = 'lpd.apps.LearnerProfileDashboardConfig'
